Refactor: Streamline the C API implementation and headers
Ultimately, webview.h should only include the C API declarations (currently called c_api.h), so that including it from any number of source files provides the bare minimum needed to use the C API only.

Consumers are expect to include either webview.h or webview.hpp directly (both of which contain the C API declarations), while the implementation, of the C API is compiled separately and added to the static library... so there's no need for this anymore.